requirements.txt
asgiref==3.8.1
attrs==25.3.0
blis==0.7.11
catalogue
certifi==2025.1.31
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.1
click==8.1.8
cryptography==44.0.2
cymem==2.0.11
dj-database-url==2.3.0
Django==4.2.20
djangorestframework==3.15.2
docx2txt==0.8
dotenv==0.9.9
idna==3.10
importlib_resources==6.4.5
joblib==1.4.2
jsonschema==4.20.0
lxml==5.3.2
murmurhash==1.0.12
nltk==3.8.1
packaging==24.2
pandas==2.1.4
pdf2image==1.17.0
pdfminer.six==20221105
pillow==10.4.0
pkgutil_resolve_name==1.3.10
plac==1.4.2
preshed==3.0.9
psycopg2-binary==2.9.10
pycparser==2.22
pycryptodome==3.22.0
pydparser
PyMuPDF==1.24.11
pyrsistent==0.20.0
pytesseract==0.3.13
python-dateutil==2.8.2
python-docx==1.1.2
python-dotenv==1.0.1
pytz==2025.2
referencing==0.35.1
regex==2024.11.6
requests==2.32.3
rpds-py==0.20.1
six==1.17.0
sortedcontainers==2.4.0
spacy==3.7.2
sqlparse==0.5.3
srsly
thinc
tqdm==4.66.1
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.2.3
wasabi==0.10.1
whitenoise==6.7.0
zipp==3.20.2
gunicorn
djangorestframework
djangorestframework-simplejwt
